requirements.txt
customtkinter 
scikit-learn 
pandas 
pdfminer.six 
docx2txt 
packaging 
pyinstaller
Pillow
requests